Rogue + addition to Character
Hey guys, I just added the Rogue class to the repository.  It's special
ability is that it can lower defense as it attacks.  Thus, I had to
adjust character.  I need to go for twenty minutes but here's this for
now.  Also, I need to update the about still, if either of you feel
especially poetic you can go for it or I'll do it in a bit.